Split to one type per file
